third_party/WebKit/LayoutTests/http/tests/notifications/request-permission-in-service-workers.html

<!doctype html>
<html>
  <head>
    <title>Notifications: Notification.requestPermission() should not be exposed in Service Workers.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
    // The Notification.requestPermission() method should not be exposed in Service Workers.
    service_worker_test(
          'resources/request-permission-worker.js',
          'Notification.requestPermission() should not be exposed in Service Workers.');
    </script>
  </body>
</html>
