third_party/WebKit/LayoutTests/platform/win/http/tests/notifications/request-permission-in-service-workers-expected.txt

This is a testharness.js-based test.
PASS Notification.requestPermission() should not be exposed in Service Workers. 
PASS Notification.requestPermission() is not exposed in Service Workers. 
Harness: the test ran to completion.
